refactor TagsCloudContainer
