+++ TagsCloudContainer/Layouting/CircularCloudLayouter.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
+using TagsCloudContainer.TagsClouds;
 using TagsCloudContainer.Visualisation;
 
 namespace TagsCloudContainer.Layouting
