+++ TagsCloudContainer/TagsCloudContainer.csproj
@@ -92,28 +92,28 @@
     <Compile Include="Formatting\IWordsFormatter.cs" />
     <Compile Include="Formatting\ToInitFormFormatter.cs" />
     <Compile Include="Formatting\ToLowerCaseFormatter.cs" />
-    <Compile Include="ITagsCloud.cs" />
-    <Compile Include="ITagsCloudFactory.cs" />
     <Compile Include="Layouting\CircularCloudLayouter.cs" />
     <Compile Include="Layouting\ITagsCloudLayouter.cs" />
     <Compile Include="Layouting\TagsCloudLayouterSettings.cs" />
     <Compile Include="Reading\DocWordsReader.cs" />
     <Compile Include="Reading\IWordsReader.cs" />
     <Compile Include="Reading\TxtWordsReader.cs" />
-    <Compile Include="TagsCloud.cs" />
-    <Compile Include="TagsCloudFactory.cs" />
-    <Compile Include="TagsCloudGeneratorSettings.cs" />
+    <Compile Include="TagsCloudGenerating\TagsCloudGenerator.cs" />
     <Compile Include="TagsCloudContainerApplication.cs" />
+    <Compile Include="TagsCloudGenerating\TagsCloudGeneratorSettings.cs" />
+    <Compile Include="TagsClouds\ITagsCloud.cs" />
+    <Compile Include="TagsClouds\ITagsCloudFactory.cs" />
+    <Compile Include="TagsClouds\TagsCloud.cs" />
+    <Compile Include="TagsClouds\TagsCloudFactory.cs" />
     <Compile Include="Tests\TagsCloudGeneratorTests.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TagsCloudGenerator.cs" />
     <Compile Include="UI\CLI.cs" />
     <Compile Include="UI\IUI.cs" />
     <Compile Include="Visualisation\ArchimedeanSpiral.cs" />
     <Compile Include="Visualisation\ImageSettings.cs" />
     <Compile Include="Visualisation\ITagsCloudRenderer.cs" />
-    <Compile Include="Visualisation\PNGTagsCloudRenderer.cs" />
+    <Compile Include="Visualisation\PngTagsCloudRenderer.cs" />
     <Compile Include="Visualisation\TagsCloudWord.cs" />
     <Compile Include="Sizing\FrequencyWordsSizer.cs" />
     <Compile Include="Sizing\IWordsSizer.cs" />
