+++ TagsCloudContainer/TagsCloudGenerating/TagsCloudGeneratorSettings.cs
